+++ bindings/python/examples/wallet/transaction_options.py
@@ -2,7 +2,7 @@ import os
 
 from dotenv import load_dotenv
 
-from iota_sdk import (RemainderValueStrategy, TaggedDataPayload,
+from iota_sdk import (RemainderValueStrategy, TaggedDataPayload, PayloadType,
                       TransactionOptions, Wallet, utf8_to_hex)
 
 load_dotenv()
@@ -27,6 +27,6 @@ params = [{
 }]
 
 transaction = account.send_with_params(params, TransactionOptions(remainder_value_strategy=RemainderValueStrategy.ReuseAddress,
-                                                                  note="my first tx", tagged_data_payload=TaggedDataPayload(utf8_to_hex("tag"), utf8_to_hex("data"))))
+                                                                  note="my first tx", tagged_data_payload=TaggedDataPayload(PayloadType.TaggedData, utf8_to_hex("tag"), utf8_to_hex("data"))))
 print(transaction)
 print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction.blockId}')