+++ bindings/python/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 1.1.3 - 202Y-MM-DD
+## 1.1.3 - 2024-02-15
 
 ### Added
 
@@ -28,10 +28,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `NodeInfoProtocol.belowMaxDepth`;
 - `{NodeInfoProtocol, RentStructure}::as_dict()`;
 
+### Changed
+
+- Made the type of `TaggedDataPayload` required as first field in `__init__()`;
+
 ### Fixed
 
 - `TransactionOptions::as_dict()` for remainder_value_strategy;
 - `Wallet::client_options` typing;
+- Made `TaggedDataPayload::{tag, data}` optional;
 
 ## 1.1.2 - 2023-12-01